backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI and ML
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
sentence-transformers==2.2.2
scikit-learn==1.3.2

# Document Processing
pdfplumber==0.9.0
pytesseract==0.3.10
python-docx==1.1.0
pypdf==3.17.0

# Excel Processing
pandas==2.1.4
openpyxl==3.1.2

# Image Processing
Pillow==10.1.0

# OCR System Requirements
# Note: Tesseract needs to be installed separately
# Windows: Download from https://github.com/UB-Mannheim/tesseract/wiki
# Mac: brew install tesseract
# Linux: sudo apt-get install tesseract-ocr

# Utilities
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.4
pathlib2==2.3.7

# Logging and Monitoring
loguru==0.7.2

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
